Use floats internally in limiter for performance reasons.

+++ src/limiter.hh
@@ -6,9 +6,9 @@
 
 class Limiter
 {
-  double ceiling        = 1;
-  double maximum        = 1;
-  double release_factor = 0;
+  float ceiling         = 1;
+  float maximum         = 1;
+  float release_factor  = 0;
   uint look_ahead       = 0;
   uint n_channels       = 0;
   uint sample_rate      = 0;
@@ -18,9 +18,9 @@ class Limiter
 public:
   Limiter (int n_channels, int sample_rate);
 
-  void set_release (double value_ms);
-  void set_attack (double value_ms);
-  void set_ceiling (double ceiling);
+  void set_release (float value_ms);
+  void set_attack (float value_ms);
+  void set_ceiling (float ceiling);
 
   std::vector<float> process (const std::vector<float>& samples);
 };
